Move economics and post-processing out of ComponentModel
ComponentModel carried three subjects: the model formulation (sets,
variables, constraints), the economic contributions and the post-processing
of a solved model. Only the first is what the class is about, and the other
two made up more than a third of the file.

Both move into a new fine/results package as mixins - ComponentEconomicsMixin
and ComponentResultsMixin - which ComponentModel inherits. Mixins rather than
collaborator objects because all seven modeling classes override the result
hooks, so a separate hierarchy would duplicate the class tree. The result
state (_rawResults, _optSummary, the optimum attributes) is initialized in
ComponentResultsMixin.__init__ and reached through the super() chain.

This is a pure move: of the 77 methods, 76 are byte-identical to their
previous version and the method surface of ComponentModel is unchanged.
__init__ is the one exception, where the result state was split off. No
signatures, no call sites and no behavior change; component.py goes from
4895 to 3035 lines.
