+++ fine/results/__init__.py
@@ -1,0 +1,6 @@
+"""Post-processing of solved component models."""
+
+from fine.results.economics import ComponentEconomicsMixin
+from fine.results.mixin import ComponentResultsMixin
+
+__all__ = ["ComponentEconomicsMixin", "ComponentResultsMixin"]
